+++ src/libopenrave/kinbodyjoint.cpp
@@ -2142,7 +2142,7 @@ void KinBody::Joint::SetMimicEquations(int iaxis, const std::string& poseq, cons
     pmimic->_posfn = CreateJointFunctionParser();
     int ret = pmimic->_posfn->ParseAndDeduceVariables(utils::SearchAndReplace(eq, poseq, jointnamepairs), resultVars);
     const size_t nVars = resultVars.size();
-    RAVELOG_DEBUG_FORMAT("Input mimic_pos = %s\nConverted into eq = %s", poseq % eq);
+    RAVELOG_VERBOSE_FORMAT("Input mimic_pos = %s\nConverted into eq = %s", poseq % eq);
     if( ret >= 0 ) {
         throw OPENRAVE_EXCEPTION_FORMAT(_("failed to set equation '%s' on %s:%s, at %d. Error is %s\n"), poseq % parent->GetName() % GetName() % ret % pmimic->_posfn->ErrorMsg(), ORE_InvalidArguments);
     }